cleanup scope referencing
Scope files are loosely coupled to tests so they should be in
same directory as test for ease of use rather than independent
scope directory.  All scopes within a test dir must be able to
satisfy the requirements for any test and vise versa.

Some naming conventions
Scope:  scope_camelCaseDescription.yml
Test:  test_camelCaseDescription.yml

And config file referencing a test and scope can also be in directory
so that editing the global config is not necessary