move build info to provider
This is important because the provider's job is to present servers
via 1-way communication.  At runtime scopes can ask for servers, but
the scope cannot tell provider which version to provide.  This may
seem like a limitation but in fact it decouples the provider from
the framework.  For example, you can use ec2 tools or ansible to install your
own version without having to read in that version from config.