Add ClusterRun Provider
Requires machine with cluster-run running
with number of nodes needed for test.

Set the config.yml to use the 'dev' provider ie:
provider: dev:10.0.0.6

Where 10.0.0.6 would be the public IP of the machine
running cluster_run.  This is needed because the docker
clients cannot reference 127.0.0.1 because that is localhost
of the container itself.