adding test: and scope: directives
Tests can now include tests and also rescope the cluster.  When
used by itself the test: directive will start another test in-line.
When the scope: directive is also present then the cluster will be
torn down to reform the new scope

This allows for the building of suites that run tests against same cluster:

-
  test: tests/simple/test_simple.yml
-
  scope: tests/simple/scope_medium.yml
  test: tests/simple/test_saveVariables.yml
-
  test: tests/simple/test_useTemplate.yml
-
  test: tests/simple/test_views.yml
-
  test: tests/simple/test_query.yml
-
  test: tests/simple/test_kvquery.yml