Streamline Testrunner workflow
Makes it much easier to run testrunner with containers.
Simply provide the -command option and this will be passed to the
testrunner container.  Also, the number of containers to start is gathered from the ini file.

./sequoia testrunner -command "-i b/resources/4-nodes-template.ini -c view-conf/py-cust.conf"

Note this change works by adding the testrunner src as a submodule into the
containers/testrunner dir.  Prior to running this then be sure to check out
the submodule:
git submodule update --init --recursive