detect when docker client is actually a swarm
This removes need to specify provider as swarm.
Which means provider can be 'file' and when client
points to a docker host that happens to be part of a swarm
then all test containers are started within the swarm and
can be ran against the provider ips.

ie:
./sequoia -client 172.23.108.94:2375 -provider file:centos_pine.yml

Where client (172.23.108.94) can be a single docker daemon or swarm
which tests the pine cluster